much better idea for hashing. need to update everything else to longs though

+++ hash.h
@@ -1,6 +1,6 @@
 #ifndef _HASH_H
 #define _HASH_H
 
-unsigned int hash(const char *str, int seed);
+unsigned long hash(const char *str, int seed);
 
 #endif
